Fix bootstrap

+++ bootstrap.sh
@@ -2,12 +2,22 @@
 
 GITHUB_ACCESS_TOKEN="a09e2e7b5488f777a79b82edd506e61ccdfcbe43"
 
-echo  "\t * Setting up TileStache"
-sudo cp ./etc/tilestache /etc/init.d
-sudo chmod 755 /etc/init.d/tilestache
-sudo /etc/init.d/tilestache start
+if [ ! -f '/home/kartverk/done_packages' ]; then
+	echo  "Setting up Tilestache"
+	sudo cp ./etc/tilestache /etc/init.d
+	sudo chmod 755 /etc/init.d/tilestache
+	sudo /etc/init.d/tilestache start
+	touch '/home/kartverk/kartverk_services/tilestache_done'
+fi
 
-echo  "\t * Setting up terrafab"
-git clone "https://$GITHUB_ACCESS_TOKEN@github.com/bengler/terrafab"
-cd terrafab
-npm install
+if [ ! -f '/home/kartverk/done_packages' ]; then
+	echo  "Setting up TerraFab"
+	git clone "https://$GITHUB_ACCESS_TOKEN@github.com/bengler/terrafab"
+	cd terrafab
+	npm install
+	cd ..
+	sudo cp ./etc/terrafab /etc/init.d
+	sudo chmod 755 /etc/init.d/terrafab
+	sudo /etc/init.d/tilestache start
+	touch '/home/kartverk/kartverk_services/tilestache_done'
+fi
